+++ conda/workflows.go
@@ -352,7 +352,9 @@ func NewEnvironment(force bool, configurations ...string) (string, error) {
 	cloud.BackgroundMetric(common.ControllerIdentity(), "rcc.env.create.start", common.Version)
 
 	lockfile := common.RobocorpLock()
+	callback := pathlib.LockWaitMessage("Serialized environment creation")
 	locker, err := pathlib.Locker(lockfile, 30000)
+	callback()
 	if err != nil {
 		common.Log("Could not get lock on live environment. Quitting!")
 		return "", err